secondsToRecordIR is now a float value. It is initialized with the duration of the room's BRIR (II). Parseval theorem verification.

+++ src/SoundSource.h
@@ -36,7 +36,7 @@ public :
 
 	void setInitialized();
 
-	void startRecordOfflineOfImpulseResponse(int _secondsToRecord);
+	void startRecordOfflineOfImpulseResponse(float _secondsToRecord);
 
 	void endRecordOfflineOfImpulseResponse();
